% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolveR-package.R
\docType{data}
\name{bardWordCount}
\alias{bardWordCount}
\title{Shakespeare word counts in the entire canon: 14,376 distinct words
appeared exactly once, 4343 words appeared twice etc.}
\usage{
data(bardWordCount)
}
\description{
Shakespeare word counts in the entire canon: 14,376 distinct words
appeared exactly once, 4343 words appeared twice etc.
}
\references{
Bradley Efron and Ronald Thisted. Estimating the number of
unseen species: How many words did Shakespeare know? Biometrika, Vol 63(3),
doi:10.1093/biomet/63.3.435.
}
\keyword{data}
