% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolveR-package.R
\docType{package}
\name{deconvolveR-package}
\alias{deconvolveR-package}
\title{R package for Empirical Bayes \eqn{g}-modeling using exponential families.}
\description{
\code{deconvolveR} is a package for Empirical Bayes Deconvolution
and Estimation. A friendly introduction is provided in the JSS
paper reference below and this package includes a vignette
containing a number of examples.
}
\references{
Bradley Efron. Empirical Bayes Deconvolution
    Estimates. Biometrika 103(1), 1-20, ISSN
    0006-3444. doi:10.1093/biomet/asv068.
    \url{http://biomet.oxfordjournals.org/content/103/1/1.full.pdf+html}

Bradley Efron and Trevor Hastie. Computer Age
    Statistical Inference.  Cambridge University Press. ISBN
    978-1-1-7-14989-2. Chapter 21.

Balasubramanian Narasimhan and Bradley
    Efron. deconvolveR: A G-Modeling Program for Deconvolution and
    Empirical Bayes Estimation. doi:10.18637/jss.v094.i11
}
