% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolveR.R
\docType{package}
\name{deconvolveR}
\alias{deconvolveR}
\alias{deconvolveR-package}
\title{R package for Empirical Bayes \eqn{g}-modeling using exponential families.
A vignette provides detailed examples and results.}
\description{
R package for Empirical Bayes \eqn{g}-modeling using exponential families.
A vignette provides detailed examples and results.
}
\references{
Bradley Efron. Empirical Bayes Deconvolution Estimates. Biometrika 103(1), 1-20,
ISSN 0006-3444. doi:10.1093/biomet/asv068.
\url{http://biomet.oxfordjournals.org/content/103/1/1.full.pdf+html}

Bradley Efron and Trevor Hastie. Computer Age Statistical Inference.
Cambridge University Press. ISBN 978-1-1-7-14989-2. Chapter 21.
}

