% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolveR.R
\docType{data}
\name{disjointTheta}
\alias{disjointTheta}
\title{A set of \eqn{\Theta} values that have a bimodal distribution for testing}
\description{
A set of \eqn{\Theta} values that have a bimodal distribution for testing
}
\keyword{data}

